Adding user registration functionality and simple email validation

+++ backend/src/main/java/com/netcracker/chargingservice/backend/repository/BillingAccountRepository.java
@@ -3,5 +3,5 @@ package com.netcracker.chargingservice.backend.repository;
 import com.netcracker.chargingservice.backend.entity.BillingAccountEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BillingAccountRepository extends CrudRepository<BillingAccountEntity, Integer> {
+public interface BillingAccountRepository extends CrudRepository<BillingAccountEntity, Long> {
 }
