+++ backend/src/main/java/com/netcracker/chargingservice/backend/repository/ContentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContentRepository extends CrudRepository<ContentEntity, Integer> {
+public interface ContentRepository extends CrudRepository<ContentEntity, Long> {
     ContentEntity findByItemName(String itemName);
 }
