+++ backend/src/main/java/com/netcracker/chargingservice/backend/repository/RolesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolesRepository extends CrudRepository<RoleEntity, Integer> {
+public interface RolesRepository extends CrudRepository<RoleEntity, Long> {
     RoleEntity findByName(String name);
 }
